Move ABI tests to compile avoidance tests
